Renamed Navbar component.
